Move subprocess command execution back into separate file
